apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.59.7
psutil==5.9.8
python-multipart==0.0.20
faster-whisper==1.1.1
httpx==0.27.2
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.155.0
apscheduler>=3.10.4
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.9.6

# ── Voice pipeline (wake word + STT + TTS) ────────────────────────────────────
openwakeword>=0.6.0
sounddevice>=0.4.6
edge-tts>=6.1.0
pygame>=2.5.0
numpy>=1.26.0
